refactor profile libraries into MobileUserProfileController
