+++ kifi-backend/modules/shoebox/app/com/keepit/controllers/mobile/MobileLibraryController.scala
@@ -362,56 +362,6 @@ class MobileLibraryController @Inject() (
       case _ => Future.successful(Forbidden)
     }
   }
-
-  def getProfileLibraries(username: Username, page: Int, pageSize: Int, filter: String) = MaybeUserAction.async { request =>
-    userCommander.userFromUsername(username) match {
-      case None =>
-        log.warn(s"unknown username ${username.value} requested")
-        Future.successful(NotFound(username.value))
-      case Some(user) =>
-        val viewer = request.userOpt
-        val paginator = Paginator(page, pageSize)
-        val imageSize = ProcessedImageSize.Large.idealSize
-        filter match {
-          case "own" =>
-            val libs = if (viewer.exists(_.id == user.id)) {
-              Json.toJson(libraryCommander.getOwnProfileLibrariesForSelf(user, paginator, imageSize).seq)
-            } else {
-              Json.toJson(libraryCommander.getOwnProfileLibraries(user, viewer, paginator, imageSize).map(LibraryCardInfo.writesWithoutOwner.writes).seq)
-            }
-            Future.successful(Ok(Json.obj("own" -> libs)))
-          case "following" =>
-            val libs = libraryCommander.getFollowingLibraries(user, viewer, paginator, imageSize).seq
-            Future.successful(Ok(Json.obj("following" -> libs)))
-          case "invited" =>
-            val libs = libraryCommander.getInvitedLibraries(user, viewer, paginator, imageSize).seq
-            Future.successful(Ok(Json.obj("invited" -> libs)))
-          case "all" if page == 0 =>
-            val ownLibsF = if (viewer.exists(_.id == user.id)) {
-              SafeFuture(Json.toJson(libraryCommander.getOwnProfileLibrariesForSelf(user, paginator, imageSize).seq))
-            } else {
-              SafeFuture(Json.toJson(libraryCommander.getOwnProfileLibraries(user, viewer, paginator, imageSize).map(LibraryCardInfo.writesWithoutOwner.writes).seq))
-            }
-            val followLibsF = SafeFuture(libraryCommander.getFollowingLibraries(user, viewer, paginator, imageSize).seq)
-            val invitedLibsF = SafeFuture(libraryCommander.getInvitedLibraries(user, viewer, paginator, imageSize).seq)
-            for {
-              ownLibs <- ownLibsF
-              followLibs <- followLibsF
-              invitedLibs <- invitedLibsF
-            } yield {
-              Ok(Json.obj(
-                "own" -> ownLibs,
-                "following" -> followLibs,
-                "invited" -> invitedLibs
-              ))
-            }
-          case "all" if page != 0 =>
-            Future.successful(BadRequest(Json.obj("error" -> "cannot_page_all_filters")))
-          case _ =>
-            Future.successful(BadRequest(Json.obj("error" -> "no_such_filter")))
-        }
-    }
-  }
 }
 
 object MobileLibraryController {
