+++ kifi-backend/modules/shoebox/app/com/keepit/controllers/mobile/MobileUserProfileController.scala
@@ -2,11 +2,13 @@ package com.keepit.controllers.mobile
 
 import com.google.inject.Inject
 import com.keepit.commanders._
+import com.keepit.common.akka.SafeFuture
 import com.keepit.common.controller.{ ShoeboxServiceController, UserActions, UserActionsHelper }
 import com.keepit.common.db.Id
 import com.keepit.common.db.slick.DBSession.RSession
 import com.keepit.common.db.slick.Database
 import com.keepit.common.social.BasicUserRepo
+import com.keepit.common.util.Paginator
 import com.keepit.model._
 import play.api.libs.json.{ JsNumber, JsObject, Json, JsValue }
 
@@ -55,6 +57,56 @@ class MobileUserProfileController @Inject() (
         )).as[JsObject]
 
         Ok(jsonFriendInfo ++ jsonProfileInfo)
+    }
+  }
+
+  def getProfileLibraries(username: Username, page: Int, pageSize: Int, filter: String) = MaybeUserAction.async { request =>
+    userCommander.userFromUsername(username) match {
+      case None =>
+        log.warn(s"unknown username ${username.value} requested")
+        Future.successful(NotFound(username.value))
+      case Some(user) =>
+        val viewer = request.userOpt
+        val paginator = Paginator(page, pageSize)
+        val imageSize = ProcessedImageSize.Large.idealSize
+        filter match {
+          case "own" =>
+            val libs = if (viewer.exists(_.id == user.id)) {
+              Json.toJson(libraryCommander.getOwnProfileLibrariesForSelf(user, paginator, imageSize).seq)
+            } else {
+              Json.toJson(libraryCommander.getOwnProfileLibraries(user, viewer, paginator, imageSize).map(LibraryCardInfo.writesWithoutOwner.writes).seq)
+            }
+            Future.successful(Ok(Json.obj("own" -> libs)))
+          case "following" =>
+            val libs = libraryCommander.getFollowingLibraries(user, viewer, paginator, imageSize).seq
+            Future.successful(Ok(Json.obj("following" -> libs)))
+          case "invited" =>
+            val libs = libraryCommander.getInvitedLibraries(user, viewer, paginator, imageSize).seq
+            Future.successful(Ok(Json.obj("invited" -> libs)))
+          case "all" if page == 0 =>
+            val ownLibsF = if (viewer.exists(_.id == user.id)) {
+              SafeFuture(Json.toJson(libraryCommander.getOwnProfileLibrariesForSelf(user, paginator, imageSize).seq))
+            } else {
+              SafeFuture(Json.toJson(libraryCommander.getOwnProfileLibraries(user, viewer, paginator, imageSize).map(LibraryCardInfo.writesWithoutOwner.writes).seq))
+            }
+            val followLibsF = SafeFuture(libraryCommander.getFollowingLibraries(user, viewer, paginator, imageSize).seq)
+            val invitedLibsF = SafeFuture(libraryCommander.getInvitedLibraries(user, viewer, paginator, imageSize).seq)
+            for {
+              ownLibs <- ownLibsF
+              followLibs <- followLibsF
+              invitedLibs <- invitedLibsF
+            } yield {
+              Ok(Json.obj(
+                "own" -> ownLibs,
+                "following" -> followLibs,
+                "invited" -> invitedLibs
+              ))
+            }
+          case "all" if page != 0 =>
+            Future.successful(BadRequest(Json.obj("error" -> "cannot_page_all_filters")))
+          case _ =>
+            Future.successful(BadRequest(Json.obj("error" -> "no_such_filter")))
+        }
     }
   }
