+++ kifi-backend/modules/shoebox/test/com/keepit/controllers/mobile/MobileLibraryControllerTest.scala
@@ -466,59 +466,6 @@ class MobileLibraryControllerTest extends Specification with ShoeboxTestInjector
       }
     }
 
-    "get profile libraries" in {
-      withDb(controllerTestModules: _*) { implicit injector =>
-        val (user1, lib1, lib2, _, _) = setupTwoUsersThreeLibraries()
-        val pubId1 = Library.publicId(lib1.id.get)(inject[PublicIdConfiguration])
-        val pubId2 = Library.publicId(lib2.id.get)(inject[PublicIdConfiguration])
-
-        val result1 = getProfileLibraries(user1, 0, 10, "own")
-        status(result1) must equalTo(OK)
-        Json.parse(contentAsString(result1)) === Json.parse(
-          s"""
-             {
-              "own" : [
-                {
-                  "id":"${pubId2.id}",
-                  "name":"Catching Jellyfish",
-                  "numFollowers":0,
-                  "numKeeps":0,
-                  "followers":[],
-                  "slug":"catching-jellyfish",
-                  "kind" : "user_created",
-                  "visibility" : "published",
-                  "numKeeps" : 0,
-                  "numFollowers" : 0,
-                  "followers": [],
-                  "lastKept": ${lib2.createdAt.getMillis},
-                  "listed": true
-                },
-                {
-                  "id":"${pubId1.id}",
-                  "name":"Krabby Patty",
-                  "numFollowers":0,
-                  "numKeeps":0,
-                  "followers":[],
-                  "slug":"krabby-patty",
-                  "kind" : "user_created",
-                  "visibility" : "secret",
-                  "numKeeps" : 0,
-                  "numFollowers" : 0,
-                  "followers": [],
-                  "lastKept": ${lib1.createdAt.getMillis},
-                  "listed": true
-                }
-              ]
-            }
-           """)
-        val result2 = getProfileLibraries(user1, 0, 10, "all")
-        status(result2) must equalTo(OK)
-        val resultJson2 = contentAsJson(result2)
-        (resultJson2 \ "own").as[Seq[JsObject]].length === 2
-        (resultJson2 \ "following").as[Seq[JsObject]].length === 0
-        (resultJson2 \ "invited").as[Seq[JsObject]].length === 0
-      }
-    }
   }
 
   private def createLibrary(user: User, body: JsObject)(implicit injector: Injector): Future[Result] = {
@@ -574,11 +521,6 @@ class MobileLibraryControllerTest extends Specification with ShoeboxTestInjector
   private def getSummariesWithUrl(user: User, body: JsObject)(implicit injector: Injector): Future[Result] = {
     inject[FakeUserActionsHelper].setUser(user)
     controller.getLibrarySummariesWithUrl()(request(routes.MobileLibraryController.getLibrarySummariesWithUrl()).withBody(body))
-  }
-
-  private def getProfileLibraries(user: User, page: Int, size: Int, filter: String)(implicit injector: Injector): Future[Result] = {
-    inject[FakeUserActionsHelper].setUser(user)
-    controller.getProfileLibraries(user.username, page, size, filter)(request(routes.MobileLibraryController.getProfileLibraries(user.username, page, size, filter)))
   }
 
   // User 'Spongebob' has one library called "Krabby Patty" (secret)
